fix: move css module files to styles directory
